[update] ラップする実装をアプリサイドで行う仕様に変更
