+++ tests/interface/app1/main.py
@@ -1,25 +1,7 @@
+from . import Metadata, urls
 from mitama.app import App
-from mitama.http import Response
-from . import Metadata
-from mitama.db.types import *
 
 meta = Metadata()
 app = App(meta)
-db = app.database()
 
-class Some(db.Model):
-    __tablename__ = 'app1_sometable'
-    id = Column(Integer, primary_key = True)
-
-db.create_all()
-
-async def home(request):
-    return Response(text = 'this is home')
-
-async def about(request):
-    return Response(text = 'this is about')
-
-app.add_routes([
-    ('/', home),
-    ('/about', about)
-])
+app.add_routes(urls.urls)
