+++ tests/interface/app1/urls.py
@@ -1,0 +1,6 @@
+from . import views
+
+urls = [
+    ('/', views.home),
+    ('/about', views.about)
+]
